src/requirements/base.txt
Django==1.11.4
django-redis==4.8.0
Pillow==4.2.1
djangorestframework==3.6.4
pytz==2017.2
elasticsearch-dsl==5.3.0
django-jet==1.0.6
django-allauth==0.33.0
django-crispy-forms==1.7.0
django-analytical==2.3.0
# Star ratings swapable-rating-model
git+https://github.com/Pustakalaya/django-star-ratings@develop
# Django hit count
git+https://github.com/Pustakalaya/django-hitcount
# Wand
Wand-0.4.4



